comms/comms-ground-control/requirements.txt
pyserial
time
openpyxl